Changed navbar to bootstrap
