+++ src/components/home/Messenger/Messenger.css
@@ -4,12 +4,12 @@
 }
 @media(min-width: 992px) {
   .messenger-card {
-    margin: 30px 40px;
+    margin: 30px 20px;
   }
 }
 @media(max-width: 992px) {
   .messenger-card {
-    margin: 30px 20px;
+    margin: 30px 40px;
   }
 }
