+++ src/components/home/NavBar/NavBar.js
@@ -1,65 +1,73 @@
 import React, {Component} from 'react';
-import AppBar from 'material-ui/AppBar';
-import {Tabs, Tab} from 'material-ui/Tabs';
+import {Image, Navbar, Nav, NavItem} from 'react-bootstrap';
 import {animateScroll as scroll, scroller} from 'react-scroll'; 
 import './NavBar.css';
-
-const scrollToComp = (name) => {
-  scroller.scrollTo(name, {offset: -120, smooth: true})
-}
-
-const NavBarTabs = 
-  <Tabs className="tabs">
-    <Tab label="About" onActive={() => scrollToComp("who-we-are")}>
-    </Tab>
-    <Tab label="Members" onActive={() => scrollToComp("brothers-are")}>
-    </Tab>
-    <Tab label="Rush" onActive={() => scrollToComp("why-rush")}>
-    </Tab>
-    <Tab label="FAQ" onActive={() => scrollToComp("faq")}>
-    </Tab>
-    <Tab label="Contact Us" onActive={() => scrollToComp("messenger")}>
-    </Tab>
-  </Tabs>
 
 class NavBar extends Component {
   scrollToTop() {
     scroll.scrollToTop();
   }
+
+  scrollToComp(name) {
+    scroller.scrollTo(name, {offset: -80, smooth: true});
+    if (document.querySelector('#hamburger').classList.contains('open')) {
+      document.querySelector('#hamburger').classList.toggle('open');
+      document.querySelector('.navbar-collapse .navbar-nav').style.opacity = 0;
+      document.querySelector('.navbar-collapse .navbar-nav').style.zIndex = -1;
+      document.querySelector('#nav-logo').style.opacity = 1;
+    }
+  }
+
   openMenu() {
     document.querySelector('#hamburger').classList.toggle('open');
-    if ( document.querySelector('#hamburger').classList.contains('open')) {
+    if (document.querySelector('#hamburger').classList.contains('open')) {
+      document.querySelector('.navbar-fixed-top').style.backgroundColor = '#000';
       document.querySelector('#nav-logo').style.opacity = 0;
-      document.querySelector('#mobile-nav').style.opacity = 1;
-      document.querySelector('#mobile-nav').style.zIndex = 100;
+      document.querySelector('.navbar-collapse .navbar-nav').style.opacity = 1;
     }
     else {
+      document.querySelector('.navbar-fixed-top').style.backgroundColor = 'rgba(10, 10, 10, .9)';
       document.querySelector('#nav-logo').style.opacity = 1;
-      document.querySelector('#mobile-nav').style.opacity = 0;
-      document.querySelector('#mobile-nav').style.zIndex = -1;
+      document.querySelector('.navbar-collapse .navbar-nav').style.opacity = 0;
     }
   }
 
   render() {
     return (
-      <AppBar
-        children={window.innerWidth > 992 && NavBarTabs}
-        style={{"position": "fixed"}}
-        iconElementLeft={
-          <a id="nav-logo" href='#' onClick={this.scrollToTop}>
-            <img className="logo" src={require('../../../../public/images/tt_logo.png')} role="presentation" />
-          </a>
-        }
-        iconElementRight={
-          <div id="hamburger" onClick={this.openMenu}>
-            <span></span>
-            <span></span>
-            <span></span>
-            <span></span>
-          </div>
-        }
-        zDepth={2}
-      />
+      <Navbar className="navbar-fixed-top"
+        inverse 
+        collapseOnSelect
+      >
+        <Navbar.Header>
+          <Navbar.Brand>
+            <a id="nav-logo" href="#" onClick={this.scrollToTop}>          
+              <Image className="logo" src={require('../../../../public/images/tt_logo.png')} />
+            </a>
+          </Navbar.Brand>
+          <Navbar.Toggle children={
+            <div id="hamburger">
+              <span></span>
+              <span></span>
+              <span></span>
+              <span></span>
+            </div>
+          }
+          onClick={this.openMenu}
+          />
+        </Navbar.Header>
+        <Nav pullLeft className="hidden-sm hidden-xs">
+            <NavItem className="navbar-title" onClick={this.scrollToTop}>Theta Tau @ UC San Diego</NavItem>
+        </Nav>
+        <Navbar.Collapse>
+          <Nav pullRight>
+            <NavItem onClick={() => this.scrollToComp("who-we-are")}>About Us</NavItem>
+            <NavItem onClick={() => this.scrollToComp("brothers-are")}>Members</NavItem>
+            <NavItem onClick={() => this.scrollToComp("why-rush")}>Rush</NavItem>
+            <NavItem onClick={() => this.scrollToComp("faq")}>FAQ</NavItem>
+            <NavItem onClick={() => this.scrollToComp("messenger")}>Contact Us</NavItem>
+          </Nav>
+        </Navbar.Collapse>
+      </Navbar>
     );
   }
 }
